Move check_congestion_jobs to gcp.py and create_congestion_job to service.py
